Corrected string escape calculation.

+++ ext/sqlite/sess_sqlite.c
@@ -142,7 +142,7 @@ PS_WRITE_FUNC(sqlite)
 	
 	t = time(NULL);
 
-	binary = emalloc(1 + 5 + vallen * (256 / 253));
+	binary = emalloc(1 + 5 + vallen * ((float) 256 / (float) 253));
 	binlen = sqlite_encode_binary((const unsigned char*)val, vallen, binary);
 	
 	rv = sqlite_exec_printf(db, "REPLACE INTO session_data VALUES('%q', '%q', %d)", NULL, NULL, &error, key, binary, t);
